rewrote it to use serenity
